requirements.txt
streamlit
pandas
openai
requests
beautifulsoup4
openpyxl
xlsxwriter
python-dotenv
